@@ -10,17 +10,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.1.0",
+    "@angular/animations": "^14.2.12",
     "@angular/cdk": "^13.3.9",
-    "@angular/common": "~13.1.0",
-    "@angular/compiler": "~13.1.0",
-    "@angular/core": "~13.1.0",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
     "@angular/fire": "^7.5.0",
-    "@angular/forms": "~13.1.0",
+    "@angular/forms": "^14.2.12",
     "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "~13.1.0",
-    "@angular/platform-browser-dynamic": "~13.1.0",
-    "@angular/router": "~13.1.0",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
     "@tailwindcss/forms": "^0.5.3",
     "daisyui": "^2.46.0",
     "firebase": "^9.15.0",
@@ -31,9 +31,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.1.3",
-    "@angular/cli": "~13.1.3",
-    "@angular/compiler-cli": "~13.1.0",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
     "@schematics/angular": "^15.0.3",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@types/jasmine": "~3.10.0",
@@ -47,6 +47,6 @@
     "karma-jasmine-html-reporter": "~1.7.0",
     "postcss": "^8.4.20",
     "tailwindcss": "^3.2.4",
-    "typescript": "~4.5.2"
+    "typescript": "~4.8.4"
   }
 }
