@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms';
 import { FirebaseService } from 'src/app/modules/auth/services/firebase/firebase.service';
 
 @Component({
@@ -9,14 +9,14 @@ import { FirebaseService } from 'src/app/modules/auth/services/firebase/firebase
 })
 export class LoginComponent implements OnInit {
 
-  formData: FormGroup;
+  formData: UntypedFormGroup;
 
   loggingIn = false;
 
   constructor(public firebase: FirebaseService) {
-    this.formData = new FormGroup({
-      'email': new FormControl('', [Validators.required, Validators.email]),
-      'password': new FormControl('', [Validators.required])
+    this.formData = new UntypedFormGroup({
+      'email': new UntypedFormControl('', [Validators.required, Validators.email]),
+      'password': new UntypedFormControl('', [Validators.required])
     });
 
    }
